# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Portfolio tracking API",
        "title": "My Portfolio API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/tax/events": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tax"
                ],
                "summary": "List everything realized in a tax year",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tax year; omit for all years",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "capital_gain | interest | dividend",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "disposal | distribution",
                        "name": "origin",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to one security",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to one account",
                        "name": "account_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to one holding",
                        "name": "holding_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 20, max 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PaginatedRealizedEvents"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tax/jurisdictions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tax"
                ],
                "summary": "List the tax jurisdictions the app knows about",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TaxJurisdiction"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tax/profiles": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tax"
                ],
                "summary": "List tax residency profiles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TaxProfile"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tax"
                ],
                "summary": "Set a user's tax residency",
                "parameters": [
                    {
                        "description": "Tax profile payload",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TaxProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TaxProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tax/recompute": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tax"
                ],
                "summary": "Re-evaluate every stored tax treatment against the current rules",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tax/rules": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tax"
                ],
                "summary": "List tax rules, optionally for one jurisdiction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter to one jurisdiction code (e.g. US-CA)",
                        "name": "jurisdiction",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TaxRule"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tax/summary": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tax"
                ],
                "summary": "Taxable income for a year, broken down per jurisdiction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tax year (defaults to the current year)",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TaxSummary"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "JurisdictionSummary": {
            "type": "object",
            "properties": {
                "buckets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/TaxBucket"
                    }
                },
                "code": {
                    "type": "string"
                },
                "excluded": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/TaxExclusion"
                    }
                },
                "level": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "taxable_total": {
                    "type": "number"
                }
            }
        },
...
        "PaginatedRealizedEvents": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RealizedEvent"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "summary": {
                    "$ref": "#/definitions/RealizedEventSummary"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "RealizedEventSummary": {
            "type": "object",
            "properties": {
                "capital_gains": {
                    "type": "number"
                },
                "dividends": {
                    "type": "number"
                },
                "interest": {
                    "type": "number"
                },
                "total": {
                    "type": "number"
                }
            }
        },
        "TaxBucket": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "character": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                }
            }
        },
        "TaxExclusion": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "label": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "TaxJurisdiction": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency_code": {
                    "type": "string"
                },
                "level": {
                    "description": "national | regional",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_code": {
                    "description": "ParentCode is the jurisdiction this one sits inside, nil for a national\njurisdiction.",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "TaxProfile": {
            "type": "object",
            "properties": {
                "country_code": {
                    "description": "CountryCode is the national jurisdiction's code; RegionCode the regional\none, nil where the country has no regional income tax or the person isn't\nsubject to one.",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "effective_from": {
                    "description": "EffectiveFrom is when this residency started. Only the current profile is\nconsulted today; storing the date means a residency history can be added\nlater without a migration.",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "region_code": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "TaxProfileRequest": {
            "type": "object",
            "required": [
                "country_code",
                "user_id"
            ],
            "properties": {
                "country_code": {
                    "type": "string"
                },
                "region_code": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "TaxRule": {
            "type": "object",
            "properties": {
                "asset_tax_class": {
                    "type": "string"
                },
                "character": {
                    "description": "Stored as tax_character: \"character\" is a type name in Postgres and can't\nbe used unquoted as a column.",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "income_type": {
                    "description": "Match columns. Empty string = matches anything.",
                    "type": "string"
                },
                "issuer_scope": {
                    "description": "see IssuerScope* above",
                    "type": "string"
                },
                "jurisdiction_code": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "priority": {
                    "description": "Priority orders the match; highest wins. Reason is the machine-readable\nlabel carried onto the resulting treatment so the UI can explain an\nexclusion (\"Excluded - U.S. Treasury interest\").",
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "term": {
                    "description": "short | long, capital gains only",
                    "type": "string"
                },
                "treatment": {
                    "description": "Outcome.",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "TaxSummary": {
            "type": "object",
            "properties": {
                "jurisdictions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/JurisdictionSummary"
                    }
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
    }
}